+++ include/openpose/utilities/keypoint.hpp
@@ -47,6 +47,21 @@ namespace op
     T getDistanceAverage(const Array<T>& keypointsA, const int personA, const Array<T>& keypointsB, const int personB,
                          const T threshold);
 
+    /**
+     * Creates and Array<T> with a specific person.
+     * @param keypoints Array<T> with the original data array to slice.
+     * @param person indicates the index of the array to extract.
+     * @param noCopy indicates whether to perform a copy. Copy will never go to undefined behavior, however, if
+     * noCopy == true, then:
+     *     1. It is faster, as no data copy is involved, but...
+     *     2. If the Array keypoints goes out of scope, then the resulting Array will provoke an undefined behavior.
+     *     3. If the returned Array is modified, the information in the Array keypoints will also be.
+     * @return Array<T> with the same dimension than keypoints expect the first dimension being 1. E.g., if keypoints
+     * is {p,k,m}, the resulting Array<T> is {1,k,m}.
+     */
+    template <typename T>
+    Array<T> getKeypointsPerson(const Array<T>& keypoints, const int person, const bool noCopy = false);
+
     template <typename T>
     float getKeypointsRoi(const Array<T>& keypoints, const int personA, const int personB, const T threshold);